Mobile: shim for BigInt added
