+++ source/native/package.json
@@ -19,6 +19,7 @@
     "@react-navigation/native-stack": "^6.5.0",
     "@react-navigation/stack": "^6.0.11",
     "@sentry/react-native": "^3.4.0",
+    "big-integer": "^1.6.51",
     "fast-text-encoding": "^1.0.3",
     "metro-react-native-babel-transformer": "^0.66.2",
     "native-base": "^3.2.2",
